refactor(sdk)!: move protocol message types to the protocol submodule
The low-level protocol message and register-input types are now importable only
from the protocol submodule (iii-sdk/protocol, iii.protocol, iii_sdk::protocol)
and are removed from the SDK root (clean break). Engine consumers repointed to
iii_sdk::protocol.
